refactor(systems): extract pure game-state functions into domain system files
Introduces lib/systems/ — pure state-transition functions for each game domain,
following a game-engine system pattern. game-context.tsx callbacks now delegate
to these pure functions instead of inlining setState logic.

- ScanSystem.ts: applyBuildScanner, applyStartScan, applyCollectScan
- MiningSystem.ts: applyMiningDone, applyRoverMiningDone
- EconomySystem.ts: applySellMinerals, applyStartRefine, applyCollectRefined, applyUpgradeLaunchpad
- ProgressionSystem.ts: applyUnlockSkillNode, applyAcceptLoan, applyAbandonMission

game-context.tsx reduced from 1015 → 846 lines. System functions are pure
(no React deps), making them independently testable. No behaviour changes.
