+++ web/lib/systems/ScanSystem.ts
@@ -1,0 +1,43 @@
+// Pure state-transition functions for the scanning station system.
+// Each function takes current GameState and returns next GameState (or same if no change).
+
+import type { GameState } from '@/lib/game-types'
+import { SCANS_PER_DAY, SCAN_DURATION_MS } from '@/lib/data'
+import { todayDateKey } from '@/lib/data'
+
+export function applyBuildScanner(s: GameState): GameState {
+  if (s.player.scannerBuilt) return s
+  return { ...s, player: { ...s.player, scannerBuilt: true } }
+}
+
+export function applyStartScan(s: GameState, targetId: string): GameState {
+  if (!s.player.scannerBuilt) return s
+  if (s.player.activeScan) return s
+  const today = todayDateKey()
+  const scanDate = s.player.scanDate ?? ''
+  const scansUsedToday = scanDate === today ? (s.player.scansUsedToday ?? 0) : 0
+  if (scansUsedToday >= SCANS_PER_DAY) return s
+  return {
+    ...s,
+    player: {
+      ...s.player,
+      activeScan: { targetId, completesAt: Date.now() + SCAN_DURATION_MS },
+      scansUsedToday: scansUsedToday + 1,
+      scanDate: today,
+    },
+  }
+}
+
+export function applyCollectScan(s: GameState): GameState {
+  const scan = s.player.activeScan
+  if (!scan || Date.now() < scan.completesAt) return s
+  const prev = s.player.targetScanCounts ?? {}
+  return {
+    ...s,
+    player: {
+      ...s.player,
+      activeScan: null,
+      targetScanCounts: { ...prev, [scan.targetId]: (prev[scan.targetId] ?? 0) + 1 },
+    },
+  }
+}
